+++ src/HasMedia/HasMedia.php
@@ -105,25 +105,71 @@ interface HasMedia
     public function registerAllMediaConversions();
 
     /**
-     * Get a collection mime types constraints validation string from its name.
+     * Get a collection mime types validation rules.
      *
      * @param string $collectionName
      *
      * @return string
+     * @deprecated Use the mimeTypesValidationRules($collectionName) method instead.
      */
     public function mimeTypesValidationConstraints(string $collectionName): string;
 
     /**
-     * Get a collection dimension validation constraints string from its name.
+     * Get a collection mime types validation rules.
      *
      * @param string $collectionName
      *
      * @return string
      */
+    public function mimeTypesValidationRules(string $collectionName): string;
+
+    /**
+     * Get a collection mimes validation rules.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     * @deprecated Use the mimesValidationRules($collectionName) method instead.
+     */
+    public function mimesValidationConstraints(string $collectionName): string;
+
+    /**
+     * Get a collection mimes validation rules.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     */
+    public function mimesValidationRules(string $collectionName): string;
+
+    /**
+     * Get a collection dimensions validation rules.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     * @deprecated Use the dimensionValidationRules($collectionName) method instead.
+     */
     public function dimensionValidationConstraints(string $collectionName): string;
 
     /**
-     * Get registered collection max width and max height from its name.
+     * Get a collection dimensions validation rules.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     */
+    public function dimensionValidationRules(string $collectionName): string;
+
+    /**
+     * Get the config max file size validation rule.
+     *
+     * @return string
+     */
+    public function sizeValidationRule(): string;
+
+    /**
+     * Get the collection declared max width and max height.
      *
      * @param string $collectionName
      *
@@ -132,40 +178,87 @@ interface HasMedia
     public function collectionMaxSizes(string $collectionName): array;
 
     /**
-     * Get the constraints validation string for a media collection.
+     * Get a collection validation rules.
+     *
+     * @param string $collectionName
+     *
+     * @return array
+     * @deprecated Use the validationRules($collectionName) method instead.
+     */
+    public function validationConstraints(string $collectionName): array;
+
+    /**
+     * Get a collection validation rules.
      *
      * @param string $collectionName
      *
      * @return array
      */
-    public function validationConstraints(string $collectionName): array;
+    public function validationRules(string $collectionName): array;
 
     /**
-     * Get the constraints legend string for a media collection.
+     * Get a collection constraints caption.
      *
      * @param string $collectionName
      *
      * @return string
+     * @deprecated Use the constraintsCaption($collectionName) method instead.
      */
     public function constraintsLegend(string $collectionName): string;
 
     /**
-     * Get a collection dimensions constraints legend string from its name.
+     * Get a collection constraints caption.
      *
      * @param string $collectionName
      *
      * @return string
+     */
+    public function constraintsCaption(string $collectionName): string;
+
+    /**
+     * Get a collection dimensions constraints caption.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     * @deprecated Use the dimensionsCaption($collectionName) method instead.
      */
     public function dimensionsLegend(string $collectionName): string;
 
     /**
-     * Get a collection mime types constraints legend string from its name.
+     * Get a collection dimensions constraints caption.
      *
      * @param string $collectionName
      *
      * @return string
      */
+    public function dimensionsCaption(string $collectionName): string;
+
+    /**
+     * Get a collection mime types constraints caption.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     * @deprecated Use the mimeTypesCaption($collectionName) method instead.
+     */
     public function mimeTypesLegend(string $collectionName): string;
+
+    /**
+     * Get a collection mime types constraints caption.
+     *
+     * @param string $collectionName
+     *
+     * @return string
+     */
+    public function mimeTypesCaption(string $collectionName): string;
+
+    /**
+     * Get the config max file size constraint caption.
+     *
+     * @return string
+     */
+    public function sizeCaption(): string;
 
     /**
      * Get declared conversions from a media collection name.