+++ tests/Unit/Caption/CollectionDimensionsCaptionTest.php
@@ -8,11 +8,9 @@ use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
 use Spatie\MediaLibrary\Tests\TestCase;
 
-class CollectionDimensionLegendTest extends TestCase
+class CollectionDimensionsCaptionTest extends TestCase
 {
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_none_when_it_is_called_with_non_existing_collection()
     {
         $testModel = new class extends TestModel
@@ -22,13 +20,11 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_none_when_it_is_called_with_non_existent_conversions()
     {
         $testModel = new class extends TestModel
@@ -38,13 +34,11 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaCollection('logo')->acceptsMimeTypes(['image/jpeg', 'image/png']);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_only_width_dimension_legend_when_only_width_is_declared()
     {
         $testModel = new class extends TestModel
@@ -59,15 +53,13 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->width(120);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals(__('medialibrary::medialibrary.constraint.dimensions.width', [
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals(__('medialibrary::medialibrary.constraints.dimensions.width.min', [
             'width' => 120,
-        ]), $dimensionsLegendString);
+        ]), $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_only_height_dimension_legend_when_only_height_is_declared()
     {
         $testModel = new class extends TestModel
@@ -82,15 +74,13 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->height(30);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals(__('medialibrary::medialibrary.constraint.dimensions.height', [
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals(__('medialibrary::medialibrary.constraints.dimensions.height.min', [
             'height' => 30,
-        ]), $dimensionsLegendString);
+        ]), $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_no_dimension_legend_when_no_size_is_declared()
     {
         $testModel = new class extends TestModel
@@ -105,13 +95,11 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb');
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_width_and_height_dimension_legend_when_both_are_declared()
     {
         $testModel = new class extends TestModel
@@ -134,16 +122,15 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 100, 70);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals(__('medialibrary::medialibrary.constraint.dimensions.both', [
-            'width'  => 100,
-            'height' => 80,
-        ]), $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals(__('medialibrary::medialibrary.constraints.dimensions.width.min', [
+                'width' => 100,
+            ]) . ' ' . __('medialibrary::medialibrary.constraints.dimensions.height.min', [
+                'height' => 80,
+            ]), $dimensionsCaptionString);
     }
 
-    /**
-     * @test
-     */
+    /** @test */
     public function it_does_not_returns_dimensions_legend_when_no_image_declared()
     {
         $testModel = new class extends TestModel
@@ -166,7 +153,7 @@ class CollectionDimensionLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 100, 70);
             }
         };
-        $dimensionsLegendString = $testModel->dimensionsLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->dimensionsCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 }