+++ tests/Unit/Rules/CollectionMimeTypesValidationRulesTest.php
@@ -2,18 +2,15 @@
 
 namespace Spatie\MediaLibrary\Tests\Unit\Extension\UrlGenerator;
 
-use Spatie\MediaLibrary\Exceptions\CollectionNotFound;
 use Spatie\MediaLibrary\File;
 use Spatie\Image\Manipulations;
 use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
 use Spatie\MediaLibrary\Tests\TestCase;
 
-class CollectionMimeTypesValidationConstraintsTest extends TestCase
+class CollectionMimeTypesValidationRulesTest extends TestCase
 {
-    /**
-     * @test
-     */
+    /** @test */
     public function it_returns_none_when_it_is_called_with_non_existing_collection()
     {
         $testModel = new class extends TestModel
@@ -23,14 +20,12 @@ class CollectionMimeTypesValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $mimeTypesValidationConstraintsString = $testModel->mimeTypesValidationConstraints('logo');
-        $this->assertEquals('', $mimeTypesValidationConstraintsString);;
+        $rules = $testModel->mimeTypesValidationRules('logo');
+        $this->assertEquals('', $rules);;
     }
 
-    /**
-     * @test
-     */
-    public function it_returns_mime_types_validation_constraints_when_declared_in_collection()
+    /** @test */
+    public function it_returns_mime_types_rules_when_declared_in_collection()
     {
         $testModel = new class extends TestModel
         {
@@ -52,14 +47,12 @@ class CollectionMimeTypesValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 100, 70);
             }
         };
-        $mimeTypesValidationConstraintsString = $testModel->mimeTypesValidationConstraints('logo');
-        $this->assertEquals('mimetypes:image/jpeg,image/png', $mimeTypesValidationConstraintsString);
+        $rules = $testModel->mimeTypesValidationRules('logo');
+        $this->assertEquals('mimetypes:image/jpeg,image/png', $rules);
     }
 
-    /**
-     * @test
-     */
-    public function it_returns_no_collection_mime_types_validation_constraints_when_none_declared()
+    /** @test */
+    public function it_returns_no_collection_mime_types_rules_when_none_declared()
     {
         $testModel = new class extends TestModel
         {
@@ -73,7 +66,7 @@ class CollectionMimeTypesValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $mimeTypesValidationConstraintsString = $testModel->mimeTypesValidationConstraints('logo');
-        $this->assertEquals('', $mimeTypesValidationConstraintsString);
+        $rules = $testModel->mimeTypesValidationRules('logo');
+        $this->assertEquals('', $rules);
     }
 }